String format validator
