+++ cel/options.go
@@ -447,6 +447,8 @@ const (
 	OptTrackCost EvalOption = 1 << iota
 
 	// OptCheckStringFormat enables compile-time checking of string.format calls for syntax/cardinality.
+	//
+	// Deprecated: use ext.ValidateFormatString() as this option is now a no-op.
 	OptCheckStringFormat EvalOption = 1 << iota
 )
